+++ dbms/src/Columns/ColumnConst.cpp
@@ -54,7 +54,7 @@ ColumnPtr ColumnConst::replicate(const Offsets & offsets) const
     return ColumnConst::create(data, replicated_size);
 }
 
-ColumnPtr ColumnConst::permute(const Permutation & perm, UInt64 limit) const
+ColumnPtr ColumnConst::permute(const Permutation & perm, size_t limit) const
 {
     if (limit == 0)
         limit = s;
